Implement asset manager
