+++ src/HeavyIron/HIAssetTypes.ts
@@ -1,0 +1,96 @@
+export const enum HIAssetType {
+    ALST = 0x414C5354,
+    ANIM = 0x414E494D,
+    ATBL = 0x4154424C,
+    ATKT = 0x41544B54,
+    BINK = 0x42494E4B,
+    BOUL = 0x424F554C,
+    BSP  = 0x42535020,
+    BUTN = 0x4255544E,
+    CAM  = 0x43414D20,
+    CCRV = 0x43435256,
+    CNTR = 0x434E5452,
+    COLL = 0x434F4C4C,
+    COND = 0x434F4E44,
+    CRDT = 0x43524454,
+    CSN  = 0x43534E20,
+    CSNM = 0x43534E4D,
+    CSSS = 0x43535353,
+    CTOC = 0x43544F43,
+    DEST = 0x44455354,
+    DPAT = 0x44504154,
+    DSCO = 0x4453434F,
+    DSTR = 0x44535452,
+    DTRK = 0x4454524B,
+    DUPC = 0x44555043,
+    DYNA = 0x44594E41,
+    EGEN = 0x4547454E,
+    ENV  = 0x454E5620,
+    FLY  = 0x464C5920,
+    FOG  = 0x464F4720,
+    GRSM = 0x4752534D,
+    GRUP = 0x47525550,
+    GUST = 0x47555354,
+    HANG = 0x48414E47,
+    JAW  = 0x4A415720,
+    JSP  = 0x4A535020,
+    LITE = 0x4C495445,
+    LKIT = 0x4C4B4954,
+    LOBM = 0x4C4F424D,
+    LODT = 0x4C4F4454,
+    MAPR = 0x4D415052,
+    MINF = 0x4D494E46,
+    MODL = 0x4D4F444C,
+    MPHT = 0x4D504854,
+    MRKR = 0x4D524B52,
+    MVPT = 0x4D565054,
+    NGMS = 0x4E474D53,
+    NPC  = 0x4E504320,
+    NPCS = 0x4E504353,
+    ONEL = 0x4F4E454C,
+    PARE = 0x50415245,
+    PARP = 0x50415250,
+    PARS = 0x50415253,
+    PEND = 0x50454E44,
+    PGRS = 0x50475253,
+    PICK = 0x5049434B,
+    PIPT = 0x50495054,
+    PKUP = 0x504B5550,
+    PLAT = 0x504C4154,
+    PLYR = 0x504C5952,
+    PORT = 0x504F5254,
+    PRJT = 0x50524A54,
+    RANM = 0x52414E4D,
+    RAW  = 0x52415720,
+    RWTX = 0x52575458,
+    SCRP = 0x53435250,
+    SDFX = 0x53444658,
+    SFX  = 0x53465820,
+    SGRP = 0x53475250,
+    SHDW = 0x53484457,
+    SHRP = 0x53485250,
+    SIMP = 0x53494D50,
+    SLID = 0x534C4944,
+    SND  = 0x534E4420,
+    SNDI = 0x534E4449,
+    SNDS = 0x534E4453,
+    SPLN = 0x53504C4E,
+    SPLP = 0x53504C50,
+    SSET = 0x53534554,
+    SUBT = 0x53554254,
+    SURF = 0x53555246,
+    TEXS = 0x54455853,
+    TEXT = 0x54455854,
+    TIMR = 0x54494D52,
+    TPIK = 0x5450494B,
+    TRIG = 0x54524947,
+    TRWT = 0x54525754,
+    UI   = 0x55492020,
+    UIFT = 0x55494654,
+    UIM  = 0x55494D20,
+    VIL  = 0x56494C20,
+    VILP = 0x56494C50,
+    VOLU = 0x564F4C55,
+    WIRE = 0x57495245,
+    ZLIN = 0x5A4C494E,
+}
